Clean up vendoring steps in plugin compiler

+++ ci/images/plugin-compiler/data/build.sh
@@ -63,16 +63,16 @@ yes | cp -r $PLUGIN_SOURCE_PATH/* $PLUGIN_BUILD_PATH || true
 cd $PLUGIN_BUILD_PATH
 
 # Handle if plugin has own vendor folder, and ignore error if not
-[ -f go.mod ] && [ ! -d ./vendor ] && GO111MODULE=on go mod vendor
-# Ensure that go modules not used
-rm -rf go.mod
+[ -f go.mod ] && [ ! -d ./vendor ] && GO111MODULE=on go mod vendor || true
 
 # We do not need to care which version of Tyk vendored in plugin, since we going to use version inside compiler
 rm -rf $PLUGIN_BUILD_PATH/vendor/github.com/TykTechnologies/tyk
 
 # Copy plugin vendored pkgs to GOPATH
-yes | cp -rf $PLUGIN_BUILD_PATH/vendor/* $GOPATH/src || true \
-        && rm -rf $PLUGIN_BUILD_PATH/vendor
+if [ -d ./vendor ]; then
+	cp -rf $PLUGIN_BUILD_PATH/vendor/* $GOPATH/src
+        rm -rf $PLUGIN_BUILD_PATH/vendor
+fi
 
 # Ensure that GW package versions have priorities
 
@@ -80,12 +80,15 @@ yes | cp -rf $PLUGIN_BUILD_PATH/vendor/* $GOPATH/src || true \
 # First we need to find which deps GW already has, remove this folders, and after copy fresh versions from GW
 
 # github.com and rest of packages have different nesting levels, so have to handle it separately
-ls -d $TYK_GW_PATH/vendor/github.com/*/* | sed "s|$TYK_GW_PATH/vendor|$GOPATH/src|g" | xargs -d '\n' rm -rf
-ls -d $TYK_GW_PATH/vendor/*/* | sed "s|$TYK_GW_PATH/vendor|$GOPATH/src|g" | grep -v github | xargs -d '\n' rm -rf
+ls -d $TYK_GW_PATH/vendor/github.com/*/* | sed "s|$TYK_GW_PATH/vendor|$GOPATH/src|g" | xargs --no-run-if-empty -d '\n' rm -rf
+ls -d $TYK_GW_PATH/vendor/*/* | sed "s|$TYK_GW_PATH/vendor|$GOPATH/src|g" | grep -v github | xargs --no-run-if-empty -d '\n' rm -rf
 
 # Copy GW dependencies
-yes | cp -rf $TYK_GW_PATH/vendor/* $GOPATH/src
-rm -rf $TYK_GW_PATH/vendor
+if [ -d $TYK_GW_PATH/vendor ]; then
+	cp -rf $TYK_GW_PATH/vendor/* $GOPATH/src
+	rm -rf $TYK_GW_PATH/vendor
+fi
+
 rm /go/src/modules.txt
 
 set +x
